requirements.txt
Acquisition==4.5
astroid==2.0.4
AuthEncoding==4.1
beautifulsoup4==4.6.3
bibliograph.core==1.0.3
bibtexparser==1.0.1
biopython==1.72
BTrees==4.5.1
certifi==2018.10.15
Chameleon==3.5
chardet==3.0.4
DateTime==4.3
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-annoying==0.10.4
django-cleanup==2.1.0
django-debug-toolbar==1.10.1
django-ipware==2.1.0
elasticsearch==6.3.1
ExtensionClass==4.4
feedparser==5.2.1
future==0.17.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.13
numpy==1.15.4
oauthlib==2.1.0
PasteDeploy==1.5.2
persistent==4.4.3
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pyparsing==2.3.0
python-gettext==3.0
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
RestrictedPython==3.6.0
RISparser==0.4.3
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sqlparse==0.2.4
transaction==2.4.0
typed-ast==1.1.0
urllib3==1.24.1
waitress==1.1.0
WebOb==1.8.4
WebTest==2.0.32
wrapt==1.10.11
WSGIProxy2==0.4.5
z3c.pt==3.1.0
zc.lockfile==1.4
ZConfig==3.3.0
zdaemon==4.3
ZEO==5.2.0
zExceptions==4.1
ZODB==5.5.1
ZODB3==3.11.0
zodbpickle==1.0.2
zope.annotation==4.7.0
zope.browser==2.3
zope.browsermenu==4.4
zope.browserpage==4.3.0
zope.browserresource==4.3
zope.cachedescriptors==4.3.1
zope.component==4.5
zope.configuration==4.3.0
zope.container==4.2.2
zope.contentprovider==4.2.1
zope.contenttype==4.4
zope.deferredimport==4.3
zope.deprecation==4.3.0
zope.dottedname==4.3
zope.event==4.4
zope.exceptions==4.3
zope.filerepresentation==4.2.0
zope.globalrequest==1.5
zope.hookable==4.2.0
zope.i18n==4.6.1
zope.i18nmessageid==4.3.1
zope.interface==4.6.0
zope.lifecycleevent==4.3
zope.location==4.2
zope.pagetemplate==4.4.1
zope.proxy==4.3.1
zope.ptresource==4.2.0
zope.publisher==5.0.1
zope.schema==4.9.3
zope.security==4.3.0
zope.sequencesort==4.1.2
zope.site==4.2.2
zope.size==4.3
zope.structuredtext==4.3
zope.tal==4.4
zope.tales==4.3
zope.testbrowser==5.3.1
zope.traversing==4.3.1
zope.viewlet==4.2
